backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pyodbc>=4.0.35
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
paho-mqtt>=1.6.1
websockets>=12.0
alembic>=1.12.0
python-dotenv>=1.0.0
